Add initial project files for desktop and wasm targets
