+++ app/src/desktopMain/kotlin/com/qamar/jetpackcomposscanner/main.kt
@@ -1,0 +1,13 @@
+package com.qamar.jetpackcomposscanner
+
+import androidx.compose.ui.window.Window
+import androidx.compose.ui.window.application
+
+fun main() = application {
+    Window(
+        onCloseRequest = ::exitApplication,
+        title = "JetpackComposScanner",
+    ) {
+        RecompositionTrackerScreen()
+    }
+}
