+++ app/src/wasmJsMain/kotlin/com/qamar/jetpackcomposscanner/main.kt
@@ -1,0 +1,13 @@
+package com.qamar.jetpackcomposscanner
+
+
+import androidx.compose.ui.ExperimentalComposeUiApi
+import androidx.compose.ui.window.ComposeViewport
+import kotlinx.browser.document
+
+@OptIn(ExperimentalComposeUiApi::class)
+fun main() {
+    ComposeViewport(document.body!!) {
+        RecompositionTrackerScreen()
+    }
+}
